pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "polars_to_ibis"
authors = [{name = "The OpenDP Project", email = "info@opendp.org"}]
readme = "README-PYPI.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dynamic = ["version", "description"]
dependencies = [
    "ibis-framework",
    "polars",
]

[project.urls]
Home = "https://github.com/opendp/polars-to-ibis"

[options]
python_requires = ">=3.10"

[project.optional-dependencies]
app = [
    "atpublic==6.0.2",
    "ibis-framework==10.8.0",
    "parsy==2.2",
    "polars-runtime-32==1.34.0",
    "polars==1.34.0",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "sqlglot==27.25.2",
    "toolz==1.0.0",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
]

[tool.pyright]
include = ["dp_wizard"]

[tool.isort]
profile = "black"
